+++ api/tour_api.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+app4=Blueprint("tour_api", __name__)
+
+@app4.route("/booking")
+def book():
+    return "123"
