+++ app.py
@@ -1,5 +1,7 @@
 from flask import *
-from api.api import app2
+from api.attractions_api import app2
+from api.user_api import app3
+from api.tour_api import app4
 
 app=Flask(__name__)
 app.config["JSON_AS_ASCII"]=False
@@ -20,8 +22,10 @@ def thankyou():
 	return render_template("thankyou.html")
 
 app.register_blueprint(app2, url_prefix="/api")
+app.register_blueprint(app3, url_prefix="/api")
+app.register_blueprint(app4, url_prefix="/api")
 
 app.debug=True
-app.run(host="0.0.0.0", port=3000)#change port to 3000
-# app.run(port=3000) #test on my computer
+# app.run(host="0.0.0.0", port=3000)#change port to 3000
+app.run(port=3000) #test on my computer
