+++ templates/index.html
@@ -24,8 +24,8 @@
         </form>
         <div class='nav_frame'>
             <div class='navigation'>
-                <h3><a href="http://3.115.234.130:3000/">台北一日遊</a></h3>
-                <!--<h3><a href="http://127.0.0.1:3000/">台北一日遊</a></h3>-->
+                <!--<h3><a href="http://3.115.234.130:3000/">台北一日遊</a></h3>-->
+                <h3><a href="http://127.0.0.1:3000/">台北一日遊</a></h3>
                 <ul>
                     <li id='greet'></li>
                     <li id='schedule'></li>
